Move common code between Session and Package to FSCollector
